feat: 스크래퍼와 API 서버 분리
