+++ TechBlogNotificationsScrapper/src/main/java/com/dgh06175/techblognotificationsscrapper/TechBlogNotificationsScrapperApplication.java
@@ -1,8 +1,13 @@
 package com.dgh06175.techblognotificationsscrapper;
 
+import com.dgh06175.techblognotificationsscrapper.controller.ScrapController;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class TechBlogNotificationsScrapperApplication {
 
@@ -10,4 +15,12 @@ public class TechBlogNotificationsScrapperApplication {
         SpringApplication.run(TechBlogNotificationsScrapperApplication.class, args);
     }
 
+    @Bean
+    CommandLineRunner runScraper(ScrapController scrapController) {
+        return args -> {
+            scrapController.scrapPosts();
+        };
+    }
 }
+
+
