+++ TechBlogNotificationsScrapper/src/main/java/com/dgh06175/techblognotificationsscrapper/repository/PostRepository.java
@@ -1,0 +1,8 @@
+package com.dgh06175.techblognotificationsscrapper.repository;
+
+import com.dgh06175.techblognotificationsscrapper.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+    Post findByLink(String link);
+}
