+++ TechBlogNotificationsScrapper/src/main/resources/application.yml
@@ -1,0 +1,10 @@
+spring:
+  application:
+    name: TechBlogNotificationsScrapper
+  jpa:
+    hibernate:
+      ddl-auto: update
+  profiles:
+    include: aws
+  main:
+    web-application-type: none
