+++ TechBlogNotificationsServer/src/main/java/com/dgh06175/techblognotificationsserver/TechBlogNotificationsServerApplication.java
@@ -1,6 +1,5 @@
 package com.dgh06175.techblognotificationsserver;
 
-import com.dgh06175.techblognotificationsserver.controller.ScrapController;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -13,12 +12,5 @@ public class TechBlogNotificationsServerApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(TechBlogNotificationsServerApplication.class, args);
-    }
-
-    @Bean
-    CommandLineRunner runScraper(ScrapController scrapController) {
-        return args -> {
-            scrapController.scrapPosts();
-        };
     }
 }
