--- TechBlogNotificationsServer/src/main/java/com/dgh06175/techblognotificationsserver/exception/ScrapHttpException.java
@@ -1,10 +1,0 @@
-package com.dgh06175.techblognotificationsserver.exception;
-
-import static com.dgh06175.techblognotificationsserver.exception.ErrorMessage.SCRAP_NULL_DOCUMENT;
-
-public class ScrapHttpException extends ScrapException {
-
-    public ScrapHttpException(String blogUrl) {
-        super(blogUrl, SCRAP_NULL_DOCUMENT);
-    }
-}
